Show categories & shorter address
Added 3 first categories of POI to dialog. To do this I also needed to
add a filter that switches underscores to spaces.
Show a shorter address that should fit better to the small version of
dialog, added a filter for the address that can take an attribute if we
want to get the shorter or the longer version.
Small finetuning of image carousel size